requirements.txt
# FirstSight: Efficient Egocentric Question Answering
# Requirements for distillation and baseline experiments

# Deep Learning Framework
torch>=2.1.0
torchvision>=0.16.0

# Transformers and Model Libraries
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.1
bitsandbytes>=0.41.3

# Data Processing
datasets>=2.15.0
pillow>=10.0.0
av>=10.0.0  # Video processing

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization and Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0

# Profiling and Monitoring
py3nvml>=0.2.7
gpustat>=1.1.0

# Optional: Logging and Experiment Tracking
tensorboard>=2.15.0
wandb>=0.16.0

# Optional: Additional optimizations
# flash-attn>=2.0.0  # Uncomment if FlashAttention is available
# ninja>=1.11.0  # For flash-attn compilation

